+++ packages/core/src/connectivity/scp-cf/xsuaa-service.ts
@@ -170,6 +170,7 @@ export async function jwtBearerTokenGrant(
 }
 
 /**
+ * Deprecated since v1.49.0. Use `fetchVerificationKeys(url: string)` instead. Credentials are ignored.
  * Fetches verification keys from the XSUAA service for the given credentials.
  * @param xsuaaCredentials - Credentials of the XSUAA service instance.
  * @param jku - Value of the jku property in the JWT header. If not provided the old legacy URL xsuaaCredentials.url/token_keys is used as a fallback which will not work for subscriber accounts created after 14th of April 2020.
@@ -181,6 +182,7 @@ export function fetchVerificationKeys(
 ): Promise<TokenKey[]>;
 
 /**
+ * Deprecated since v1.49.0. Use `fetchVerificationKeys(url: string)` instead. Credentials are ignored.
  * Fetches verification keys from the XSUAA service for the given URL, with the given pair of credentials.
  * @param url - URL of the XSUAA service instance.
  * @param clientId - Client ID of the XSUAA service instance.
@@ -193,9 +195,17 @@ export function fetchVerificationKeys(
   clientSecret: string
 ): Promise<TokenKey[]>;
 
+/**
+ * Fetches verification keys from the XSUAA service for the given URL.
+ * @param url - URL of the XSUAA service instance.
+ * @returns An array of token keys.
+ */
+export function fetchVerificationKeys(url: string): Promise<TokenKey[]>;
+
 export function fetchVerificationKeys(
   xsuaaUriOrCredentials: string | XsuaaServiceCredentials,
   clientIdOrJku?: string,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   clientSecret?: string
 ): Promise<TokenKey[]> {
   // The case where the XsuaaServiceCredentials are given as object
@@ -207,45 +217,29 @@ export function fetchVerificationKeys(
           'Please provide the right URL given by the field `jku` in the JWT header.'
       );
       return executeFetchVerificationKeys(
-        `${xsuaaUriOrCredentials.url}/token_keys`,
-        xsuaaUriOrCredentials.clientid,
-        xsuaaUriOrCredentials.clientsecret
+        `${xsuaaUriOrCredentials.url}/token_keys`
       );
     }
 
     return executeFetchVerificationKeys(clientIdOrJku);
   }
   // The three strings case
-  return executeFetchVerificationKeys(
-    xsuaaUriOrCredentials,
-    clientIdOrJku,
-    clientSecret
-  );
+  return executeFetchVerificationKeys(xsuaaUriOrCredentials);
 }
 
-function executeFetchVerificationKeys(
-  url: string,
-  clientId?: string,
-  clientSecret?: string
-): Promise<TokenKey[]> {
+// TODO: in v2 move this implementation to `fetchVerificationKeys`
+function executeFetchVerificationKeys(url: string): Promise<TokenKey[]> {
   const config: AxiosRequestConfig = {
     url,
     method: 'GET'
   };
-  if (clientId && clientSecret) {
-    const authHeader = headerForClientCredentials({
-      username: clientId,
-      password: clientSecret
-    });
-    config.headers = { Authorization: authHeader };
-  }
 
   return axios
     .request(config)
     .then(resp => resp.data.keys.map(k => renameKeys(tokenKeyKeyMapping, k)))
     .catch(error => {
       throw new ErrorWithCause(
-        `Failed to fetch verification keys from XSUAA service instance ${url}!`,
+        `Failed to fetch verification keys from XSUAA service instance "${url}".`,
         error
       );
     });