+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ## Compatibility Notes
 
--
+- [core] Do not use client credentials to fetch public keys for token verification.
 
 ## New Functionality
 
@@ -27,7 +27,6 @@
 ## Fixed Issues
 
 - [core] Fix verification key retrieval to match the given key id (`kid`).
-
 
 # 1.48.1